requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
python-jose==3.3.0
Pillow==10.2.0
transformers==4.37.2